Attempt at fixit the DayTimeEditor

+++ toolkit/day_time_editor/curve_widget.py
@@ -211,26 +211,26 @@ class CurveWidget(QWidget):
         # Draw vertical lines
         painter.setPen(QColor(200, 200, 200))
         for i in range(num_vert_lines + 1):
-            line_pos = i * line_spacing_x + self._legend_border - 1
-            painter.drawLine(line_pos, self._bar_h, line_pos, canvas_height + self._bar_h)
+            line_pos = int(i * line_spacing_x + self._legend_border - 1)
+            painter.drawLine(line_pos, self._bar_h, line_pos, int(canvas_height + self._bar_h))
 
         # Draw horizontal lines
         painter.setPen(QColor(200, 200, 200))
         for i in range(num_horiz_lines):
-            line_pos = canvas_height - i * line_spacing_y + self._bar_h
+            line_pos = int(canvas_height - i * line_spacing_y + self._bar_h)
             painter.drawLine(self._legend_border, line_pos, self.width(), line_pos)
 
         # Draw vetical legend labels
         painter.setPen(QColor(120, 120, 120))
         for i in range(num_horiz_lines):
-            line_pos = canvas_height - i * line_spacing_y + self._bar_h
+            line_pos = int(canvas_height - i * line_spacing_y + self._bar_h)
             # painter.drawText(6, line_pos + 3, str(round(float(i) / (num_horiz_lines-1), 2)))
             painter.drawText(
                 6, line_pos + 3, self._unit_processor(float(i) / (num_horiz_lines - 1)))
 
         # Draw horizontal legend labels
         for i in range(num_vert_lines + 1):
-            line_pos = i * line_spacing_x + self._legend_border
+            line_pos = int(i * line_spacing_x + self._legend_border)
             offpos_x = -14
             if i == 0:
                 offpos_x = -2
@@ -245,7 +245,7 @@ class CurveWidget(QWidget):
             last_value = 0
             for i in range(canvas_width):
                 rel_offset = i / (canvas_width - 1.0)
-                curve_height = self._get_y_value_for(curve.get_value(rel_offset))
+                curve_height = int(self._get_y_value_for(curve.get_value(rel_offset)))
 
                 if i == 0:
                     last_value = curve_height
@@ -258,8 +258,8 @@ class CurveWidget(QWidget):
             painter.setBrush(QColor(240, 240, 240))
 
             for cv_index, (x, y) in enumerate(curve.control_points):
-                offs_x = x * canvas_width + self._legend_border
-                offs_y = (1 - y) * canvas_height + self._bar_h
+                offs_x = int(x * canvas_width + self._legend_border)
+                offs_y = int((1 - y) * canvas_height + self._bar_h)
 
                 if (self._selected_point and self._selected_point[0] == index and
                    self._selected_point[1] == cv_index):
@@ -308,8 +308,8 @@ class CurveWidget(QWidget):
         if self._drag_point:
             painter.setBrush(QColor(200, 200, 200))
             painter.setPen(QColor(90, 90, 90))
-            offs_x = max(0, min(
-                canvas_width + 10, self._drag_time * canvas_width + self._legend_border - 19))
+            offs_x = int(max(0, min(
+                canvas_width + 10, self._drag_time * canvas_width + self._legend_border - 19)))
             offs_y = self.height() - self._legend_border
             minutes = int(self._drag_time * 24 * 60)
             hours = minutes / 60
@@ -327,7 +327,7 @@ class CurveWidget(QWidget):
         pen.setStyle(Qt.DashLine)
         painter.setPen(pen)
 
-        xoffs = self._legend_border + self._current_time * (canvas_width - 1)
+        xoffs = int(self._legend_border + self._current_time * (canvas_width - 1))
         painter.drawLine(xoffs, self._bar_h, xoffs, self._bar_h + canvas_height)
 
         # Draw usage hints
